fix role

+++ src/pages/DetailProject/DetailProject.js
@@ -1,4 +1,4 @@
-import React, { useContext, useState } from 'react';
+import React, { useContext, useEffect, useState } from 'react';
 import classNames from 'classnames/bind';
 import style from './DetailProject.module.scss';
 import Button from '~/component/Buttton/Button';
@@ -12,24 +12,46 @@ import { UserContext } from '~/contexts/user/userContext';
 import WorkService from '~/services/work/workServices';
 import { ProjectContext } from '~/contexts/project/projectContext';
 import ModalAccept from '~/component/ModalAccept/ModalAccept';
+import { Tooltip } from 'react-tooltip';
 const cx = classNames.bind(style);
 function DetailProject() {
    const navigate = useNavigate();
-   const { loadingDetailsProject, parseuser } = useContext(UserContext);
-   const { detailProject } = useContext(ProjectContext);
+   const { loadingDetailsProject, parseuser, dataUserProfile } = useContext(UserContext);
+   const { detailProject, members, setMembers } = useContext(ProjectContext);
    const [toggle, setToggle] = useState(false);
    const [toggleMoveToTrash, setToggleMoveToTrash] = useState(false);
+   const [isLoading, setIsLoading] = useState(false);
 
    const params = useParams();
    const projectService = new WorkService();
 
+   useEffect(() => {
+      getMembers();
+   }, [detailProject]);
+
    // 3. Func
    const handleMoveToTrash = async (id) => {
+      if (isLoading) return;
+      setIsLoading(true);
       const moveToTrash = await projectService.deleteProject(id, parseuser?._id);
       if (moveToTrash.status === 200) {
          navigate('/project');
       }
+      setIsLoading(false);
    };
+
+   // Get Member
+   const getMembers = async () => {
+      if (detailProject.codeProject) {
+         const listMembers = await projectService.getMember({ codeProject: detailProject?.codeProject });
+         if (listMembers.status === 200) setMembers(listMembers.data);
+      }
+   };
+
+   const roleUsers = members?.filter((user) => {
+      return user._id === dataUserProfile._id;
+   });
+   const roleUser = roleUsers[0];
 
    return (
       <div className={cx('wrapper')}>
@@ -66,14 +88,39 @@ function DetailProject() {
                            </div>
                            <div className={cx('details-button')}>
                               <div onClick={() => setToggle(true)}>
-                                 <Button backgroundNone leftIcon={<MenuIcon />} style={{ height: '32px' }}></Button>
+                                 <Button
+                                    data-tooltip-id="delete-project"
+                                    data-tooltip-content="You are not an admin."
+                                    data-tooltip-place="top"
+                                    backgroundNone
+                                    leftIcon={<MenuIcon />}
+                                    disable={roleUser?.role !== 'admin'}
+                                    style={{
+                                       cursor: roleUser?.role === 'admin' ? 'pointer' : 'not-allowed',
+                                       height: '32px',
+                                       background: 'var(--ds-background-neutral, rgba(9, 30, 66, 0.04))',
+                                    }}
+                                 ></Button>
                               </div>
+                              {roleUser?.role !== 'admin' && (
+                                 <Tooltip
+                                    id="delete-project"
+                                    style={{
+                                       backgroundColor: 'var(--ds-background-neutral-bold, #44546f)',
+                                       color: 'var(--ds-text-inverse, #FFFFFF)',
+                                       padding: 'var(--ds-space-025, 2px) var(--ds-space-075, 6px)',
+                                       fontSize: 'var(--ds-font-size-075, 12px)',
+                                       maxWidth: '240px',
+                                       textAlign: 'center',
+                                    }}
+                                 />
+                              )}
                               <MenuProject
                                  onClick={() => {
                                     setToggleMoveToTrash(true);
                                     setToggle(false);
                                  }}
-                                 isOpen={toggle}
+                                 isOpen={roleUser?.role === 'admin' && toggle}
                                  onClose={() => setToggle(false)}
                               />
                            </div>
@@ -91,7 +138,7 @@ function DetailProject() {
                         />
                      )}
                      <div className={cx('details-form')}>
-                        <FormChangeProject id={params?.id} />
+                        <FormChangeProject id={params?.id} roleUser={roleUser} />
                      </div>
                   </div>
                </div>
